Adding completion as a subset of the tools command
